requirements.txt
pandas
scikit-learn
torch
torchvision
transformers
timm
lightgbm
Pillow
tqdm
optuna
joblib